Move fns make_id, reconstruct_key to DataKey
