+++ crates/kas-core/src/model/filter/filter_list.rs
@@ -121,12 +121,6 @@ impl<T: ListData, F: Filter<T::Item> + SingleData> ListData for FilteredList<T, 
     fn len(&self) -> usize {
         self.view.borrow().1.len()
     }
-    fn make_id(&self, parent: &WidgetId, key: &Self::Key) -> WidgetId {
-        self.data.make_id(parent, key)
-    }
-    fn reconstruct_key(&self, parent: &WidgetId, child: &WidgetId) -> Option<Self::Key> {
-        self.data.reconstruct_key(parent, child)
-    }
 
     fn iter_from(&self, start: usize, limit: usize) -> Self::KeyIter<'_> {
         let end = self.len().min(start + limit);
