+++ crates/kas-view/src/matrix_view.rs
@@ -10,7 +10,7 @@ use kas::event::components::ScrollComponent;
 use kas::event::{Command, Scroll};
 use kas::layout::{solve_size_rules, AlignHints};
 #[allow(unused)] use kas::model::SharedData;
-use kas::model::{MatrixData, SharedDataMut};
+use kas::model::{DataKey, MatrixData, SharedDataMut};
 use kas::prelude::*;
 use kas::theme::SelectionStyle;
 #[allow(unused)] // doc links
@@ -378,7 +378,7 @@ impl_scope! {
                     let ci = first_col + cn;
                     let i = solver.data_to_child(ci, ri);
                     let key = T::make_key(col, &row);
-                    let id = self.data.make_id(self.id_ref(), &key);
+                    let id = key.make_id(self.id_ref());
                     let w = &mut self.widgets[i];
                     if w.key.as_ref() != Some(&key) {
                         // Reset widgets to ensure input state such as cursor
@@ -468,7 +468,7 @@ impl_scope! {
             })
         }
         fn find_child_index(&self, id: &WidgetId) -> Option<usize> {
-            let key = self.data.reconstruct_key(self.id_ref(), id);
+            let key = T::Key::reconstruct_key(self.id_ref(), id);
             if key.is_some() {
                 self.widgets
                     .iter()
@@ -749,7 +749,7 @@ impl_scope! {
                             let key = T::make_key(&ck, &rk);
                             assert_eq!(
                                 self.widgets[index].widget.id(),
-                                self.data.make_id(self.id_ref(), &key),
+                                key.make_id(self.id_ref()),
                             );
                         }
