+++ examples/data-list-view.rs
@@ -109,12 +109,6 @@ impl ListData for MySharedData {
     fn len(&self) -> usize {
         self.data.borrow().len
     }
-    fn make_id(&self, parent: &WidgetId, key: &Self::Key) -> WidgetId {
-        parent.make_child(*key)
-    }
-    fn reconstruct_key(&self, parent: &WidgetId, child: &WidgetId) -> Option<Self::Key> {
-        child.next_key_after(parent)
-    }
 
     fn iter_from(&self, start: usize, limit: usize) -> Self::KeyIter<'_> {
         let len = self.len();
